requirements.txt
decorator==4.0.10
olefile==0.44
Pillow==4.0.0
praw==4.2.0
prawcore==0.6.0
requests==2.12.4
six==1.10.0
update-checker==0.15
